+++ src/controller/map/simplexml/node/DjinnNode.java
@@ -27,7 +27,7 @@ public class DjinnNode extends AbstractFixtureNode<Djinn> {
 	 */
 	@Override
 	public Djinn produce(final PlayerCollection players, final Warning warner) throws SPFormatException {
-		final Djinn fix = new Djinn(Long.parseLong(getProperty("id")));
+		final Djinn fix = new Djinn(Integer.parseInt(getProperty("id")));
 		if (hasProperty("file")) {
 			fix.setFile(getProperty("file"));
 		}
