+++ src/controller/map/simplexml/node/OasisNode.java
@@ -27,7 +27,7 @@ public class OasisNode extends AbstractFixtureNode<Oasis> {
 	 */
 	@Override
 	public Oasis produce(final PlayerCollection players, final Warning warner) throws SPFormatException {
-		final Oasis fix = new Oasis(Long.parseLong(getProperty("id")));
+		final Oasis fix = new Oasis(Integer.parseInt(getProperty("id")));
 		if (hasProperty("file")) {
 			fix.setFile(getProperty("file"));
 		}
