+++ src/model/map/fixtures/Hill.java
@@ -15,7 +15,7 @@ public class Hill implements TerrainFixture, HasImage {
 	/**
 	 * @param idNum the ID number.
 	 */
-	public Hill(final long idNum) {
+	public Hill(final int idNum) {
 		id = idNum;
 	}
 	/**
@@ -60,7 +60,7 @@ public class Hill implements TerrainFixture, HasImage {
 	 */
 	@Override
 	public int hashCode() {
-		return (int) id;
+		return id;
 	}
 	/**
 	 * @param fix
@@ -75,12 +75,12 @@ public class Hill implements TerrainFixture, HasImage {
 	/**
 	 * ID number.
 	 */
-	private final long id; // NOPMD
+	private final int id; // NOPMD
 	/**
 	 * @return a UID for the fixture.
 	 */
 	@Override
-	public long getID() {
+	public int getID() {
 		return id;
 	}
 	/**
